删除 macOS 应用支持
