+++ .github/ISSUE_TEMPLATE/ui-small-change.yml
@@ -7,7 +7,7 @@ body:
   - type: markdown
     attributes:
       value: |
-        请用中文描述这个 UI 小改动。提交前请尽量确认它在 iOS 和 macOS 上都符合预期。
+        请用中文描述这个 UI 小改动。提交前请尽量确认它在 iOS 上符合预期。
 
   - type: textarea
     id: background
@@ -38,9 +38,8 @@ body:
       description: 描述改完之后用户应该看到什么、能做什么。
       placeholder: |
         例如：
-        1. macOS：下载进度应单独打开一个页面/窗口，避免阻塞主界面操作。
-        2. iOS：下载进度可以收起到底部区域，用户点击后再展开查看详情。
-        3. 显示明确进度条，最好包含已完成比例、当前步骤或下载状态。
+        1. 下载进度可以收起到底部区域，用户点击后再展开查看详情。
+        2. 显示明确进度条，最好包含已完成比例、当前步骤或下载状态。
     validations:
       required: true
 
@@ -61,8 +60,7 @@ body:
       description: 用可勾选清单描述这个 issue 完成时需要满足的条件。
       value: |
         - [ ] iOS 上已验证展示和交互符合预期。
-        - [ ] macOS 上已验证展示和交互符合预期。
-        - [ ] 文案、间距、颜色和平台默认行为没有明显不一致。
+        - [ ] 文案、间距、颜色和 iOS 默认行为没有明显不一致。
     validations:
       required: true
