+++ .github/workflows/ci.yml
@@ -92,10 +92,8 @@ jobs:
       - name: Run tests
         run: |
           ./Scripts/validate_data_schemas.sh
+          ./Scripts/generate_xcodeproj.sh
           ./Scripts/test.sh
 
-      - name: Generate Xcode project
-        run: ./Scripts/generate_xcodeproj.sh
-
-      - name: Build iOS and macOS apps
+      - name: Build iOS app
         run: ./Scripts/build_apps.sh
