--- Apps/macOSApp/README.md
@@ -1,3 +1,0 @@
-# macOS App
-
-SwiftUI macOS application target for browsing the Chinese calendar.
