+++ Makefile
@@ -7,12 +7,12 @@ help:
 	@echo "  make xcodeproj             Generate ChineseCalendar.xcodeproj"
 	@echo "  make buildserver           Generate VS Code buildServer.json"
 	@echo "  make setup                 Generate Xcode project and build server config"
-	@echo "  make test                  Run Swift package tests"
+	@echo "  make test                  Run tests on iOS Simulator"
 	@echo "  make format                Format Swift sources"
 	@echo "  make format-check          Check Swift formatting"
 	@echo "  make lint                  Run SwiftLint"
 	@echo "  make ci                    Run local CI checks"
-	@echo "  make build-apps            Build iOS and macOS app targets"
+	@echo "  make build-apps            Build the iOS app target"
 	@echo "  make seed-store            Build/update the versioned SwiftData seed store artifact"
 	@echo "  make validate-data-schemas Validate generated data schema artifacts"
