+++ Scripts/ci.sh
@@ -5,6 +5,6 @@ set -euo pipefail
 ./Scripts/format.sh --check
 ./Scripts/lint.sh
 ./Scripts/validate_data_schemas.sh
-./Scripts/test.sh
 ./Scripts/generate_xcodeproj.sh
+./Scripts/test.sh
 ./Scripts/build_apps.sh
