+++ Scripts/fix_xcodegen_local_package_backlinks.py
@@ -27,6 +27,7 @@ IOS_TEST_TARGETS = [
     "ChineseCalendarCoreTests",
     "ChineseCalendarDataTests",
     "ChineseCalendarLoggingTests",
+    "NavigationCoreTests",
     "ChineseCalendarUITests",
 ]
