+++ Sources/NavigationCore/NavigationPresentationModifiers.swift
@@ -1,16 +1,11 @@
 import SwiftUI
 
 public extension View {
-    @ViewBuilder
     func navigationFullScreenCover<Item: Identifiable>(
         item: Binding<Item?>,
         onDismiss: (() -> Void)? = nil,
         @ViewBuilder content: @escaping (Item) -> some View
     ) -> some View {
-        #if os(iOS)
-            fullScreenCover(item: item, onDismiss: onDismiss, content: content)
-        #else
-            sheet(item: item, onDismiss: onDismiss, content: content)
-        #endif
+        fullScreenCover(item: item, onDismiss: onDismiss, content: content)
     }
 }
