+++ Sources/NavigationCore/README.md
@@ -21,7 +21,7 @@
 - `NavigationRequestResult`：说明请求已立即执行，还是需要等待 presentation dismissal 完成。
 - `NavigationPresentationNode<Destination>`：表示一个拥有独立 push path 的 presentation，允许继续嵌套 sheet/full-screen。
 - `NavigationPresentationNodeView`：为 presentation 创建独立 `NavigationStack`，使用调用方提供的 destination builder 渲染 root 和 pushed destination，并递归承载子 presentation。
-- `navigationFullScreenCover`：iOS 使用 `fullScreenCover`，macOS 退化为 `sheet`。
+- `navigationFullScreenCover`：使用 `fullScreenCover` 承载全屏 presentation。
 
 ## 基本用法
